Rename Icon Button component to Button
