+++ components/ui/index.js
@@ -1,4 +1,4 @@
 export {default as Card} from './Card/Card';
 export {default as DefaultText} from './DefaultText/DefaultText';
-export {default as IconButton} from './IconButton/IconButton';
+export {default as Button} from './Button/Button';
 export {default as ProfilePicture} from './ProfilePicture/ProfilePicture';
