+++ screens/Contacts/Contacts.js
@@ -2,16 +2,13 @@ import React, {useLayoutEffect} from 'react';
 import {View} from 'react-native';
 
 import {ContactList} from '../../components/Contacts';
-import {IconButton} from '../../components/ui';
+import {Button} from '../../components/ui';
 import styles from './Contacts.styles';
 
 const setNavOptions = navigation => {
   navigation.setOptions({
     headerRight: () => (
-      <IconButton
-        name="add"
-        onPress={() => navigation.navigate('ManageContact')}
-      />
+      <Button name="add" onPress={() => navigation.navigate('ManageContact')} />
     ),
   });
 };
